fix add sprint button
